Add files via upload
This is just a test board that I made. It has the basics of a chessboard: checkerboard, row and column labels. If we need to make changes or use a different aspect of tkinter that is fine we can just manipulate the code to use whatever we need this can be the basics of the pattern layout and labels.